delete button

+++ server/controllers/posts.js
@@ -44,7 +44,11 @@ export const updatePost = async (req, res) => {
 };
 
 export const deletePost = async (req, res) => {
-  const { id } = req.params;
+  const { id: _id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(_id)) {
+    return res.status(404).send("No post with that id");
+  }
 
   try {
     const data = await PostMessage.findByIdAndDelete(id);
